feat: refactor code login anonymous
