+++ test/element/left-side-bar/left-side-bar.ts
@@ -1,7 +1,15 @@
-class LeftSideBar {
+class LeftSidebar {
     get logo() {
         return $('#appbar-logo');
     }
+
+    get appbarProfile() {
+        return $('#appbar-profile');
+    }
+
+    async clickAppbarProfile() {
+        return this.appbarProfile.click();
+    }
 }
 
-export default new LeftSideBar();
+export default new LeftSidebar();
