+++ test/pageobjects/account/myprofile/myprofile.ts
@@ -1,0 +1,7 @@
+class MyProfile {
+    get displayName() {
+        return $('#profile-content-display-name');
+    }
+}
+
+export default new MyProfile();
