+++ test/pageobjects/page.ts
@@ -2,4 +2,7 @@ export default class Page {
     open(path: string) {
         return browser.url(`${path}`);
     }
+    async getAllWindows() {
+        return browser.getWindowHandles();
+    }
 }
